fix issue with SSR
